+++ src/OpenTelemetry/Trace/ActivityExporter.cs
@@ -14,122 +14,14 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Diagnostics;
-using System.Threading;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Trace
 {
     /// <summary>
-    /// Enumeration used to define the result of an export operation.
-    /// </summary>
-    public enum ExportResult
-    {
-        /// <summary>
-        /// Export succeeded.
-        /// </summary>
-        Success = 0,
-
-        /// <summary>
-        /// Export failed.
-        /// </summary>
-        Failure = 1,
-    }
-
-    /// <summary>
     /// Activity exporter base class.
     /// </summary>
-    public abstract class ActivityExporter : IDisposable
+    public abstract class ActivityExporter : BaseExporter<Activity>
     {
-        private int shutdownCount;
-
-        /// <summary>
-        /// Exports a batch of <see cref="Activity"/> objects.
-        /// </summary>
-        /// <param name="batch">Batch of <see cref="Activity"/> objects to export.</param>
-        /// <returns>Result of the export operation.</returns>
-        public abstract ExportResult Export(in Batch<Activity> batch);
-
-        /// <summary>
-        /// Attempts to shutdown the exporter, blocks the current thread until
-        /// shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <exception cref="System.ArgumentOutOfRangeException">
-        /// Thrown when the <c>timeoutMilliseconds</c> is smaller than -1.
-        /// </exception>
-        /// <remarks>
-        /// This function guarantees thread-safety. Only the first call will
-        /// win, subsequent calls will be no-op.
-        /// </remarks>
-        public bool Shutdown(int timeoutMilliseconds = Timeout.Infinite)
-        {
-            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-            {
-                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds));
-            }
-
-            if (Interlocked.Increment(ref this.shutdownCount) > 1)
-            {
-                return false; // shutdown already called
-            }
-
-            try
-            {
-                return this.OnShutdown(timeoutMilliseconds);
-            }
-            catch (Exception ex)
-            {
-                OpenTelemetrySdkEventSource.Log.SpanProcessorException(nameof(this.Shutdown), ex);
-                return false;
-            }
-        }
-
-        /// <inheritdoc/>
-        public void Dispose()
-        {
-            this.Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Called by <c>Shutdown</c>. This function should block the current
-        /// thread until shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <remarks>
-        /// This function is called synchronously on the thread which made the
-        /// first call to <c>Shutdown</c>. This function should not throw
-        /// exceptions.
-        /// </remarks>
-        protected virtual bool OnShutdown(int timeoutMilliseconds)
-        {
-            return true;
-        }
-
-        /// <summary>
-        /// Releases the unmanaged resources used by this class and optionally
-        /// releases the managed resources.
-        /// </summary>
-        /// <param name="disposing">
-        /// <see langword="true"/> to release both managed and unmanaged resources;
-        /// <see langword="false"/> to release only unmanaged resources.
-        /// </param>
-        protected virtual void Dispose(bool disposing)
-        {
-        }
     }
 }
